+++ src/config.rs
@@ -29,9 +29,10 @@ pub struct ConfigItem {
     pub subtemplate: Option<String>,
     pub hook: Option<String>,
     pub rewrite: Option<bool>,
-    pub light: Option<bool>,
+    pub lightweight: Option<bool>,
     pub start: Option<String>,
     pub end: Option<String>,
+    pub light: Option<Box<ConfigItem>>,
 }
 
 impl Config {